Melhorando a organização de pastas e arquivos
