+++ src/styles/global.css
@@ -1,0 +1,54 @@
+@import url('https://fonts.googleapis.com/css2?family=Inter:wght@100..900&display=swap');
+@import url('https://fonts.cdnfonts.com/css/poetsen-one');
+
+
+*{
+    margin: 0;
+    padding: 0;
+}
+
+html{
+    font-size: 62.5%; 
+    
+    
+}
+
+input{
+    width: 330px;
+    height: 45px;
+    padding-left: 15px;
+    box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
+    border-radius: 13.5px;
+    border: 2px solid var(--Chambray-300, #A0C6F0);
+    background: var(--Star-Dust-50, #F6F6F6);
+    font-family: 'Poetsen One', sans-serif;
+    color: black;
+    margin-bottom: 15px;
+}
+
+input:focus {
+    outline: none; 
+}
+
+.btn{
+    width: 345px;
+    height: 45px;
+    align-self: center;
+    border-radius: 13.5px;
+    border: none;
+    background: var(--Chambray-500, #5387DE);
+    box-shadow: 0px 3.6px 3.6px 0px rgba(255, 255, 255, 0.50) inset, 0px -3.6px 3.6px 0px rgba(0, 0, 0, 0.50) inset, 0px 0px 0px 0.9px rgba(0, 0, 0, 0.25);
+    color: #FFF;
+    font-size: 14.4px;
+    font-family: 'Poetsen One', sans-serif;
+   
+}
+
+.btn:hover{
+    background-color: var(--Chambray-500, #6593e0);
+}
+
+
+.btn:focus{
+    outline: none;
+}
