+++ src/styles/login.css
@@ -1,0 +1,62 @@
+
+.container{
+    display: flex;
+    flex-direction: row;
+    height: 100vh;
+    width: 100vw;
+    background-color: #FFF;
+    
+}
+
+.loginContainer{
+    display: flex;
+    flex-direction: column;
+    margin: auto;
+    
+    
+}
+
+.formContainer{
+    display: flex;
+    flex-direction: column;
+    
+}
+
+.title{
+    font-family: 'Inter', sans-serif;
+    font-weight: 400;
+    color: #000;
+    text-align: center;
+    margin-bottom: 80px;
+    font-size: 31.5px;
+}
+
+.svg-input{
+    position: absolute;
+    background-size: cover;
+    width: 20px;
+    height: 20px;
+    top: 57%;
+    right: 38%;
+    background-image: url(../assets/invisible-2.svg);
+    background-repeat: no-repeat;
+    
+}
+
+.esqueci{
+    text-align: right;
+    color: #6D6D6D;
+    text-decoration: none;
+    margin-bottom: 20px;
+    font-family: 'Poetsen One', sans-serif;
+    font-size: 14.4px;
+
+}
+.esqueci:hover{
+    color: #6D6D6D;
+}
+
+
+
+
+
